move useStoreLocator to hooks
