+++ packages/template-retail-react-app/app/components/store-locator-modal/index.test.jsx
@@ -6,14 +6,10 @@
  */
 
 import React from 'react'
-import PropTypes from 'prop-types'
-import StoreLocatorModal, {
-    useStoreLocator
-} from '@salesforce/retail-react-app/app/components/store-locator-modal/index'
+import StoreLocatorModal from '@salesforce/retail-react-app/app/components/store-locator-modal/index'
 import {renderWithProviders} from '@salesforce/retail-react-app/app/utils/test-utils'
 import {act, fireEvent, screen} from '@testing-library/react'
 import {rest} from 'msw'
-import mockConfig from '@salesforce/retail-react-app/config/mocks/default'
 
 const mockStoresData = [
     {
@@ -239,47 +235,5 @@ describe('StoreLocatorModal', () => {
             expect(screen.getByText(store.name)).toBeInTheDocument()
             expect(screen.getByText(store.address1)).toBeInTheDocument()
         })
-    })
-})
-
-describe('useStoreLocator', () => {
-    const MockComponent = ({store} = {}) => {
-        const {selectedStore, setStore, isStoreSelected} = useStoreLocator()
-
-        return (
-            <>
-                <div data-testid="selected-store">{JSON.stringify(selectedStore)}</div>
-                <div data-testid="is-store-selected">{isStoreSelected.toString()}</div>
-                <button data-testid="set-store" onClick={() => setStore(store)} />
-            </>
-        )
-    }
-    MockComponent.propTypes = {
-        store: PropTypes.object
-    }
-
-    test('no store selected', () => {
-        renderWithProviders(<MockComponent />)
-        expect(screen.getByTestId('selected-store')).toHaveTextContent(/{}/i)
-        expect(screen.getByTestId('is-store-selected')).toHaveTextContent(/false/i)
-    })
-
-    test('store selected', () => {
-        const store = {
-            id: 'test-store',
-            name: 'Test Store',
-            inventoryId: 'inventory'
-        }
-
-        const expectedStoreData = JSON.stringify(store)
-
-        renderWithProviders(<MockComponent store={store} />)
-        fireEvent.click(screen.getByTestId('set-store'))
-
-        expect(screen.getByTestId('selected-store')).toHaveTextContent(expectedStoreData)
-        expect(screen.getByTestId('is-store-selected')).toHaveTextContent(/true/i)
-        expect(localStorage.getItem(`store_${mockConfig.app.defaultSite}`)).toEqual(
-            expectedStoreData
-        )
     })
 })
