+++ packages/template-retail-react-app/app/components/store-locator-modal/store-locator-content.test.jsx
@@ -13,7 +13,7 @@ import PropTypes from 'prop-types'
 import {STORE_LOCATOR_NUM_STORES_PER_LOAD} from '@salesforce/retail-react-app/app/constants'
 import {rest} from 'msw'
 import {StoreLocatorContext} from '@salesforce/retail-react-app/app/components/store-locator-modal/index'
-import {useStoreLocator} from '@salesforce/retail-react-app/app/components/store-locator-modal/index'
+import {useStoreLocator} from '@salesforce/retail-react-app/app/hooks/use-store-locator'
 const mockStoresData = [
     {
         address1: '162 University Ave',
